5-add-reference-and-tag-model: rename common model to Entity
