+++ api/src/models/Entity/operator/types.ts
